feat: Update storyblokMiddleware
