--- packages/middleware-chain/src/middlewareChain/index.ts
@@ -1,1 +1,0 @@
-export * from './middlewareChain';
