+++ packages/storyblok-preview/src/index.ts
@@ -1,3 +1,4 @@
 export * from './components/DraftModeProvider/DraftModeProvider';
 export * from './utils/isDraftMode/isDraftMode';
 export * from './actions/actions';
+export * from './storyblok-middleware';
